split out mapping-options too
