+++ frontend/src/metabase/lib/click-behavior.js
@@ -9,10 +9,8 @@ import Question from "metabase-lib/lib/Question";
 import { TemplateTagVariable } from "metabase-lib/lib/Variable";
 import { TemplateTagDimension } from "metabase-lib/lib/Dimension";
 import { isa, TYPE } from "metabase/lib/types";
-import {
-  dimensionFilterForParameter,
-  variableFilterForParameter,
-} from "metabase/meta/Parameter";
+import { variableFilterForParameter } from "metabase/meta/Parameter";
+import { dimensionFilterForParameter } from "metabase/parameters/utils/field-filters";
 
 export function getDataFromClicked({
   extraData: { dashboard, parameterValuesBySlug, userAttributes } = {},
