+++ frontend/test/metabase/meta/Parameter.unit.spec.js
@@ -1,15 +1,11 @@
-import _ from "underscore";
-
 import MetabaseSettings from "metabase/lib/settings";
 import {
-  dimensionFilterForParameter,
   getTagOperatorFilterForParameter,
   getParameterTargetField,
   dateParameterValueToMBQL,
   stringParameterValueToMBQL,
   numberParameterValueToMBQL,
   parameterToMBQLFilter,
-  parameterOptionsForField,
   normalizeParameterValue,
   deriveFieldOperatorFromParameter,
   getTemplateTagParameters,
@@ -305,182 +301,6 @@ describe("metabase/meta/Parameter", () => {
           metadata,
         ),
       ).toEqual(["between", ["field", PRODUCTS.RATING.id, null], 1, 100]);
-    });
-  });
-
-  describe("parameterOptionsForField", () => {
-    const field = {
-      isDate: () => false,
-      isID: () => false,
-      isCategory: () => false,
-      isCity: () => false,
-      isState: () => false,
-      isZipCode: () => false,
-      isCountry: () => false,
-      isNumber: () => false,
-      isString: () => false,
-      isLocation: () => false,
-    };
-    it("should return relevantly typed options for date field", () => {
-      const dateField = {
-        ...field,
-        isDate: () => true,
-      };
-      const availableOptions = parameterOptionsForField(dateField);
-      expect(
-        availableOptions.length > 0 &&
-          availableOptions.every(option => option.type.startsWith("date")),
-      ).toBe(true);
-    });
-
-    it("should return relevantly typed options for id field", () => {
-      const idField = {
-        ...field,
-        isID: () => true,
-      };
-      const availableOptions = parameterOptionsForField(idField);
-      expect(
-        availableOptions.length > 0 &&
-          availableOptions.every(option => option.type.startsWith("id")),
-      ).toBe(true);
-    });
-
-    it("should return the specific location/state option for a state field", () => {
-      const stateField = {
-        ...field,
-        isState: () => true,
-      };
-      const availableOptions = parameterOptionsForField(stateField);
-      expect(availableOptions).toEqual([
-        expect.objectContaining({ type: "location/state" }),
-      ]);
-    });
-
-    it("as a result of all location parameters haiving subtypes should return nothing for a generic location field", () => {
-      const locationField = { ...field, isLocation: () => true };
-      const availableOptions = parameterOptionsForField(locationField);
-      expect(availableOptions).toEqual([]);
-    });
-  });
-
-  describe("dimensionFilterForParameter", () => {
-    const field = {
-      isDate: () => false,
-      isID: () => false,
-      isCategory: () => false,
-      isCity: () => false,
-      isState: () => false,
-      isZipCode: () => false,
-      isCountry: () => false,
-      isNumber: () => false,
-      isString: () => false,
-      isLocation: () => false,
-    };
-    const typelessDimension = {
-      field: () => field,
-    };
-
-    [
-      [
-        { type: "date/single" },
-        {
-          type: "date",
-          field: () => ({ ...field, isDate: () => true }),
-        },
-      ],
-      [
-        { type: "id" },
-        {
-          type: "id",
-          field: () => ({ ...field, isID: () => true }),
-        },
-      ],
-      [
-        { type: "category" },
-        {
-          type: "category",
-          field: () => ({ ...field, isCategory: () => true }),
-        },
-      ],
-      [
-        { type: "location/city" },
-        {
-          type: "city",
-          field: () => ({ ...field, isCity: () => true }),
-        },
-      ],
-      [
-        { type: "number/!=" },
-        {
-          type: "number",
-          field: () => ({
-            ...field,
-            isNumber: () => true,
-            isCoordinate: () => false,
-          }),
-        },
-      ],
-      [
-        { type: "string/=" },
-        {
-          type: "category",
-          field: () => ({
-            ...field,
-            isCategory: () => true,
-          }),
-        },
-      ],
-      [
-        { type: "string/!=" },
-        {
-          type: "category",
-          field: () => ({
-            ...field,
-            isCategory: () => true,
-          }),
-        },
-      ],
-      [
-        { type: "string/starts-with" },
-        {
-          type: "string",
-          field: () => ({
-            ...field,
-            isString: () => true,
-          }),
-        },
-      ],
-    ].forEach(([parameter, dimension]) => {
-      it(`should return a predicate that evaluates to true for a ${dimension.type} dimension when given a ${parameter.type} parameter`, () => {
-        const predicate = dimensionFilterForParameter(parameter);
-        expect(predicate(typelessDimension)).toBe(false);
-        expect(predicate(dimension)).toBe(true);
-      });
-    });
-
-    it("should return a predicate that evaluates to false for a coordinate dimension when given a number parameter", () => {
-      const coordinateDimension = {
-        field: () => ({
-          ...field,
-          isNumber: () => true,
-          isCoordinate: () => true,
-        }),
-      };
-
-      const predicate = dimensionFilterForParameter({ type: "number/between" });
-      expect(predicate(coordinateDimension)).toBe(false);
-    });
-
-    it("should return a predicate that evaluates to false for a location dimension when given a category parameter", () => {
-      const locationDimension = {
-        field: () => ({
-          ...field,
-          isLocation: () => true,
-        }),
-      };
-
-      const predicate = dimensionFilterForParameter({ type: "category" });
-      expect(predicate(locationDimension)).toBe(false);
     });
   });
